+++ packages/plugin-console-breadcrumbs/test/console-breadcrumbs.test.js
@@ -24,10 +24,10 @@ describe('plugin: console breadcrumbs', () => {
       }
     })
     expect(c.breadcrumbs.length).toBe(3)
-    expect(c.breadcrumbs[0].metaData['[0]']).toBe('check 1, 2')
-    expect(c.breadcrumbs[1].metaData['[0]']).toBe('null')
-    expect(c.breadcrumbs[2].metaData['[0]']).toBe('{"foo":[1,2,3,"four"]}')
-    expect(c.breadcrumbs[2].metaData['[1]']).toBe('{"pets":{"cat":"scratcher","dog":"pupper","rabbit":"sniffer"}}')
+    expect(c.breadcrumbs[0].metadata['[0]']).toBe('check 1, 2')
+    expect(c.breadcrumbs[1].metadata['[0]']).toBe('null')
+    expect(c.breadcrumbs[2].metadata['[0]']).toBe('{"foo":[1,2,3,"four"]}')
+    expect(c.breadcrumbs[2].metadata['[1]']).toBe('{"pets":{"cat":"scratcher","dog":"pupper","rabbit":"sniffer"}}')
     // undo the global side effects of wrapping console.* for the rest of the tests
     plugin.destroy()
   })
@@ -39,14 +39,14 @@ describe('plugin: console breadcrumbs', () => {
     c.use(plugin)
     expect(() => console.log(Object.create(null))).not.toThrow()
     expect(c.breadcrumbs.length).toBe(1)
-    expect(c.breadcrumbs[0].name).toBe('Console output')
-    expect(c.breadcrumbs[0].metaData['[0]']).toBe('[Unknown value]')
+    expect(c.breadcrumbs[0].message).toBe('Console output')
+    expect(c.breadcrumbs[0].metadata['[0]']).toBe('[Unknown value]')
     plugin.destroy()
   })
 
-  it('should not be enabled when autoBreadcrumbs=false', () => {
+  it('should not be enabled when enabledBreadcrumbTypes=[]', () => {
     const c = new Client(VALID_NOTIFIER)
-    c.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa', autoBreadcrumbs: false })
+    c.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa', enabledBreadcrumbTypes: [] })
     c.configure()
     c.use(plugin)
     console.log(123)
@@ -54,9 +54,9 @@ describe('plugin: console breadcrumbs', () => {
     plugin.destroy()
   })
 
-  it('should be enabled when autoBreadcrumbs=false, consoleBreadcrumbsEnabled=true', () => {
+  it('should be enabled when enabledBreadcrumbTypes=["log"]', () => {
     const c = new Client(VALID_NOTIFIER)
-    c.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa', autoBreadcrumbs: false, consoleBreadcrumbsEnabled: true })
+    c.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa', enabledBreadcrumbTypes: ['log'] })
     c.configure()
     c.use(plugin)
     console.log(123)
@@ -71,16 +71,6 @@ describe('plugin: console breadcrumbs', () => {
     c.use(plugin)
     console.log(123)
     expect(c.breadcrumbs.length).toBe(0)
-    plugin.destroy()
-  })
-
-  it('can be enabled when releaseStage=development', () => {
-    const c = new Client(VALID_NOTIFIER)
-    c.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa', releaseStage: 'development', consoleBreadcrumbsEnabled: true })
-    c.configure()
-    c.use(plugin)
-    console.log(123)
-    expect(c.breadcrumbs.length).toBe(1)
     plugin.destroy()
   })
 })