+++ packages/plugin-navigation-breadcrumbs/test/navigation-breadcrumbs.test.js
@@ -23,20 +23,20 @@ describe('plugin: navigation breadcrumbs', () => {
 
     // first ensure that the pushState command works to change the url of the page
     window.history.replaceState(state, 'bar', 'network-breadcrumb-test.html')
-    expect(c.breadcrumbs[c.breadcrumbs.length - 1].metaData.to).toMatch(/^\/?network-breadcrumb-test\.html$/)
+    expect(c.breadcrumbs[c.breadcrumbs.length - 1].metadata.to).toMatch(/^\/?network-breadcrumb-test\.html$/)
 
     window.history.replaceState(state, 'bar')
     // then ensure that it works with `undefined` as the url parameter (IE11-specific issue)
-    expect(c.breadcrumbs[c.breadcrumbs.length - 1].metaData.to).toMatch(/^\/?network-breadcrumb-test\.html$/)
+    expect(c.breadcrumbs[c.breadcrumbs.length - 1].metadata.to).toMatch(/^\/?network-breadcrumb-test\.html$/)
 
     expect(c.breadcrumbs.length).toBe(6)
 
     done()
   })
 
-  it('should not be enabled when autoBreadcrumbs=false', () => {
+  it('should not be enabled when enabledBreadcrumbTypes=[]', () => {
     const c = new Client(VALID_NOTIFIER)
-    c.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa', autoBreadcrumbs: false })
+    c.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa', enabledBreadcrumbTypes: [] })
     c.configure()
     const { winHandlers, docHandlers, window } = getMockWindow()
     c.use(plugin, window)
@@ -81,9 +81,9 @@ describe('plugin: navigation breadcrumbs', () => {
     setTimeout(() => done(), 1)
   })
 
-  it('should be enabled when autoBreadcrumbs=false and navigationBreadcrumbsEnabled=true', () => {
+  it('should be enabled when enabledReleaseStages=["navigation"]', () => {
     const c = new Client(VALID_NOTIFIER)
-    c.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa', autoBreadcrumbs: false, navigationBreadcrumbsEnabled: true })
+    c.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa', enabledReleaseStages: ['navigation'] })
     c.configure()
     const { winHandlers, docHandlers, window } = getMockWindow()
     c.use(plugin, window)