+++ packages/plugin-react-native-app-state-breadcrumbs/test/app-state.test.js
@@ -28,17 +28,17 @@ describe('plugin: react native app state breadcrumbs', () => {
     _cb('background')
     expect(client.breadcrumbs.length).toBe(1)
     expect(client.breadcrumbs[0].type).toBe('state')
-    expect(client.breadcrumbs[0].name).toBe('App state changed')
-    expect(client.breadcrumbs[0].metaData).toEqual({ state: 'background' })
+    expect(client.breadcrumbs[0].message).toBe('App state changed')
+    expect(client.breadcrumbs[0].metadata).toEqual({ state: 'background' })
 
     _cb('active')
     expect(client.breadcrumbs.length).toBe(2)
     expect(client.breadcrumbs[1].type).toBe('state')
-    expect(client.breadcrumbs[1].name).toBe('App state changed')
-    expect(client.breadcrumbs[1].metaData).toEqual({ state: 'active' })
+    expect(client.breadcrumbs[1].message).toBe('App state changed')
+    expect(client.breadcrumbs[1].metadata).toEqual({ state: 'active' })
   })
 
-  it('should not be enabled when autoBreadcrumbs=false', () => {
+  it('should not be enabled when enabledBreadcrumbTypes=null', () => {
     let _cb
     const AppState = {
       addEventListener: (type, fn) => {
@@ -50,14 +50,14 @@ describe('plugin: react native app state breadcrumbs', () => {
     })
 
     const client = new Client(VALID_NOTIFIER)
-    client.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa', autoBreadcrumbs: false })
+    client.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa', enabledBreadcrumbTypes: null })
     client.configure()
     client.use(plugin)
 
     expect(_cb).toBe(undefined)
   })
 
-  it('should not be enabled when appStateBreadcrumbsEnabled=false', () => {
+  it('should not be enabled when enabledBreadcrumbTypes=[]', () => {
     let _cb
     const AppState = {
       addEventListener: (type, fn) => {
@@ -69,14 +69,14 @@ describe('plugin: react native app state breadcrumbs', () => {
     })
 
     const client = new Client(VALID_NOTIFIER)
-    client.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa', appStateBreadcrumbsEnabled: false })
+    client.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa', enabledBreadcrumbTypes: [] })
     client.configure()
     client.use(plugin)
 
     expect(_cb).toBe(undefined)
   })
 
-  it('should be enabled when autoBreadcrumbs=false and appStateBreadcrumbsEnabled=true', () => {
+  it('should be enabled when enabledBreadcrumbTypes=["state"]', () => {
     let _cb
     const AppState = {
       addEventListener: (type, fn) => {
@@ -88,7 +88,7 @@ describe('plugin: react native app state breadcrumbs', () => {
     })
 
     const client = new Client(VALID_NOTIFIER)
-    client.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa', autoBreadcrumbs: false, appStateBreadcrumbsEnabled: true })
+    client.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa', enabledBreadcrumbTypes: ['state'] })
     client.configure()
     client.use(plugin)