+++ packages/plugin-react-native-orientation-breadcrumbs/test/orientation.test.js
@@ -33,8 +33,8 @@ describe('plugin: react native orientation breadcrumbs', () => {
     currentDimensions = { height: 200, width: 100 }
     _cb()
     expect(client.breadcrumbs.length).toBe(1)
-    expect(client.breadcrumbs[0].name).toBe('Orientation changed')
-    expect(client.breadcrumbs[0].metaData).toEqual({ from: 'landscape', to: 'portrait' })
+    expect(client.breadcrumbs[0].message).toBe('Orientation changed')
+    expect(client.breadcrumbs[0].metadata).toEqual({ from: 'landscape', to: 'portrait' })
 
     currentDimensions = { height: 200, width: 100 }
     _cb()
@@ -43,11 +43,11 @@ describe('plugin: react native orientation breadcrumbs', () => {
     currentDimensions = { height: 100, width: 200 }
     _cb()
     expect(client.breadcrumbs.length).toBe(2)
-    expect(client.breadcrumbs[1].name).toBe('Orientation changed')
-    expect(client.breadcrumbs[1].metaData).toEqual({ from: 'portrait', to: 'landscape' })
+    expect(client.breadcrumbs[1].message).toBe('Orientation changed')
+    expect(client.breadcrumbs[1].metadata).toEqual({ from: 'portrait', to: 'landscape' })
   })
 
-  it('should not be enabled when autoBreadcrumbs=false', () => {
+  it('should not be enabled when enabledBreadcrumbTypes=null', () => {
     let _cb
     const Dimensions = {
       addEventListener: (type, fn) => {
@@ -59,14 +59,14 @@ describe('plugin: react native orientation breadcrumbs', () => {
     })
 
     const client = new Client(VALID_NOTIFIER)
-    client.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa', autoBreadcrumbs: false })
+    client.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa', enabledBreadcrumbTypes: null })
     client.configure()
     client.use(plugin)
 
     expect(_cb).toBe(undefined)
   })
 
-  it('should not be enabled when orientationBreadcrumbsEnabled=false', () => {
+  it('should not be enabled when enabledBreadcrumbTypes=[]', () => {
     let _cb
     const Dimensions = {
       addEventListener: (type, fn) => {
@@ -78,14 +78,14 @@ describe('plugin: react native orientation breadcrumbs', () => {
     })
 
     const client = new Client(VALID_NOTIFIER)
-    client.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa', orientationBreadcrumbsEnabled: false })
+    client.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa', enabledBreadcrumbTypes: [] })
     client.configure()
     client.use(plugin)
 
     expect(_cb).toBe(undefined)
   })
 
-  it('should be enabled when autoBreadcrumbs=false and orientationBreadcrumbsEnabled=true', () => {
+  it('should be enabled when enabledBreadcrumbTypes=["state"]', () => {
     let _cb
     const Dimensions = {
       addEventListener: (type, fn) => {
@@ -98,7 +98,7 @@ describe('plugin: react native orientation breadcrumbs', () => {
     })
 
     const client = new Client(VALID_NOTIFIER)
-    client.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa', autoBreadcrumbs: false, orientationBreadcrumbsEnabled: true })
+    client.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa', enabledBreadcrumbTypes: ['state'] })
     client.configure()
     client.use(plugin)